refactor(wordle): alerts
