+++ src/pages/wordle/components/PlayAgain.jsx
@@ -1,0 +1,17 @@
+import styled from "@emotion/styled";
+import { useContext } from "react";
+import { Context, playAgain } from "../context";
+
+const Button = styled.span`
+  background-color: var(--background);
+  border-radius: 0.5rem;
+  color: var(--foreground);
+  cursor: pointer;
+  padding: 0.5rem 1rem;
+`;
+
+export const PlayAgain = () => {
+  const { dispatch } = useContext(Context);
+
+  return <Button onClick={() => dispatch(playAgain())}>Play again!</Button>;
+};
